// webpack собирает, но много багов еще по коду
