+++ data/cart.json
@@ -1,1 +1,1 @@
-[]
+[{"id_product":4,"product_name":"Наушники","price":450,"count":2}]
