+++ src/data/cart.json
@@ -1,0 +1,1 @@
+[{"id_product":11,"product_name":"Лампа","price":3500,"count":1},{"id_product":10,"product_name":"Кресло","price":10000,"count":1},{"id_product":9,"product_name":"Наушники2","price":450,"count":1}]
