+++ src/index.html
@@ -11,6 +11,6 @@
 <body>
 <div id="app"></div>
 <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
-<script src="dist/main.js"></script>
+<script src="dist/build.js"></script>
 </body>
 </html>
