+++ src/main.js
@@ -1,0 +1,9 @@
+import Vue from 'vue';
+import App from './app.vue';
+
+window.addEventListener('load', function () {
+    new Vue({
+        el: '#app',
+        render: h => h(App)
+    })
+})
