+++ src/services/storage.services.js
@@ -1,0 +1,17 @@
+class StorageService {
+    storage = sessionStorage;
+
+    getItem(key) {
+        return this.storage.getItem(key);
+    }
+
+    setItem(key, value) {
+        return this.storage.setItem(key, typeof value === 'object' ? JSON.stringify(value) : value);
+    }
+
+    removeItem(key) {
+        return this.storage.removeItem(key);
+    }
+}
+
+export default new StorageService();
