+++ style/style.css
@@ -288,9 +288,11 @@ header, footer {
 }
 
 .query-not-found {
+    margin-top: 50px;
+    margin-bottom: 50px;
     display: block;
     text-align: center;
     color:#666666;
-    font-size: 20px;
+    font-size: 25px;
     font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
 }
