+++ webpack.config.js
@@ -28,9 +28,7 @@ module.exports = {
               loader: "babel-loader",
               options: {
                   presets: ['@babel/preset-env'],
-                  plugins: [["@babel/plugin-proposal-class-properties"], [
-                    "@babel/plugin-transform-runtime",
-                    { "regenerator": true }]
+                  plugins: [["@babel/plugin-proposal-class-properties"], ["@babel/plugin-transform-runtime", { "regenerator": true }]
                 ]
               },
             }
